todo_core/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["evildes <latsplaym@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
markdown = "^3.7"
drf-yasg = "^1.21.7"
setuptools = "^74.1.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
